Restructure and update appledevicetable.md
Removed iPad model detection results.
Separated the tables for zoomed and standard mode mobile detection.

The changes to the tables required updates to the anchor names.

Reordered the groupings be high to low by percentage and new to old release date for the standard mode mobile table. These changes will be replicated in the zoomed mode and desktop mode tables before releasing.